requirements.txt
newspaper4k @ git+https://github.com/AndyTheFactory/newspaper4k/@c5e4170918a6d1e99cb1bab6fd188ee8ed5a2afa # forked version of newspaper4k with some fixes
html2text==2024.2.26
requests==2.32.3
feedgen==1.0.0
sqlalchemy>=2.0.28
flask-restx==1.3.0
APScheduler==3.11.0
waitress==3.0.2
markdownify==0.14.1
ftfy==6.3.1